refactor: create separate file for moveFiles function
